EEM119/Final_Project/azure_web_app/requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
azure-storage-blob==2.0.1
pandas==0.24.2
